requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.5
requests==2.31.0
beautifulsoup4==4.12.2
python-multipart==0.0.6 